# openapi.yaml
openapi: 3.1.0
info:
  title: Mammography Second Opinion API
  version: 0.1.0
  description: FFDM mammography standalone kernel API.
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId, correlationId]
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
            correlationId:
              type: string
...
    CaseResponse:
      type: object
      required: [caseId, status, assessment, qc, generation, review, delivery, safety]
      properties:
        caseId:
          type: string
          format: uuid
        status:
          type: string
          enum: [Submitted, AwaitingReview, Finalized]
        assessment:
          anyOf:
            - $ref: '#/components/schemas/CaseAssessment'
            - type: 'null'
        qc:
          anyOf:
            - $ref: '#/components/schemas/CaseQuality'
            - type: 'null'
        generation:
          anyOf:
            - $ref: '#/components/schemas/DraftGenerationSummary'
            - type: 'null'
        review:
          anyOf:
            - $ref: '#/components/schemas/CaseReview'
            - type: 'null'
        delivery:
          anyOf:
            - $ref: '#/components/schemas/CaseDelivery'
            - type: 'null'
        safety:
          $ref: '#/components/schemas/CaseSafety'
    CaseReportResponse:
      type: object
      required: [caseId, status, report]
      properties:
        caseId:
          type: string
          format: uuid
        status:
          type: string
          enum: [Finalized]
        report:
          $ref: '#/components/schemas/RenderedReport'
...
    CaseListResponse:
      type: object
      required: [cases, total, limit, offset]
      properties:
        cases:
          type: array
          items:
            $ref: '#/components/schemas/CaseSummary'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
paths:
...
  /api/v1/integration-seams/python-sidecar:
    get:
      operationId: renderPythonSidecarIntegrationSeam
      responses:
        '200':
          description: Live integration seam for the optional Python sidecar scaffold
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PythonSidecarIntegrationSeamResponse'
        '500':
          description: Unexpected server-side failure during sidecar seam rendering
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/cases:
    get:
      operationId: listCases
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Paginated list of mammography case summaries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseListResponse'
        '500':
          description: Unexpected server-side failure during case listing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      operationId: createCase
      responses:
        '201':
          description: Draft mammography case created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Unexpected server-side failure during case intake
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/cases/{caseId}/review:
    post:
      operationId: finalizeCaseReview
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reviewerName, reviewerRole, disposition, finalBiradsCategory, finalSummary, reviewNotes]
              properties:
                reviewerName:
                  type: string
                reviewerRole:
                  type: string
                disposition:
                  type: string
                  enum: [confirmed, modified]
                finalBiradsCategory:
                  type: string
                  enum: ['0', '1', '2', '3', '4', '5', '6']
                finalSummary:
                  type: string
                reviewNotes:
                  type: string
      responses:
        '200':
          description: Clinician review finalized successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseResponse'
        '400':
          description: Invalid clinician review body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Case not found in the configured persistence store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Case review cannot be finalized from the current workflow state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Unexpected server-side failure during review finalization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/cases/{caseId}/report:
    get:
      operationId: renderCaseReport
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Finalized clinician report rendered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseReportResponse'
        '404':
          description: Case not found in the configured persistence store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Case exists but clinician finalization has not completed yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Unexpected server-side failure during report rendering
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/cases/{caseId}/report/export:
    get:
      operationId: exportCaseReport
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Finalized clinician report exported as a plain-text attachment
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: Case not found in the configured persistence store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Case exists but clinician finalization has not completed yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Unexpected server-side failure during report export
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /api/v1/cases/{caseId}/review-seams/ohif:
    get:
      operationId: renderOhifReviewSeam
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OHIF-compatible launch manifest for case review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OhifReviewSeamResponse'
        '404':
          description: Case not found in the configured persistence store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Unexpected server-side failure during OHIF seam rendering
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/cases/{caseId}/archive-seams/dicomweb:
    get:
      operationId: renderDicomwebArchiveSeam
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Orthanc-compatible DICOMweb archive handoff manifest for the case
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DicomwebArchiveSeamResponse'
        '404':
          description: Case not found in the configured persistence store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Unexpected server-side failure during DICOMweb archive seam rendering
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/cases/{caseId}:
    get:
      operationId: getCaseById
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Persisted draft mammography case retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseResponse'
        '404':
          description: Case not found in the configured persistence store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Unexpected server-side failure during case retrieval
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/cases/{caseId}/events:
    get:
      operationId: getCaseEventsById
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Persisted lifecycle events for the mammography case
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseEventsResponse'
        '404':
          description: Case not found in the configured persistence store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Unexpected server-side failure during case event retrieval
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/cases/{caseId}/report/seal:
    post:
      operationId: sealCaseReport
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportSealInput'
      responses:
        '201':
          description: Report integrity seal created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportSealResponse'
        '400':
          description: Invalid seal input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Case not found in the configured persistence store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Case is not ready for sealing or has already been sealed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Unexpected server-side failure during report sealing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...